various win32 patches/enhancements from Ludwig Ertl

+++ src/mwin/winlib/progbar.c
@@ -206,7 +206,7 @@ void pbarOnDraw (HWND hwnd, HDC hdc, PROGRESSDATA* pData, BOOL fVertical,
 	    cr = GetSysColor(COLOR_BTNSHADOW);
 	FillBox (hdc, x + ((w - maxw)>>1), y + ((h - charh) > 1), maxw,
 		charh - 1, cr);
-	FillBox (hdc, x, y, (int)((int32_t)w*d/100L), h, BLUE);
+	FillBox (hdc, x, y, (int)((LONG)w*d/100L), h, BLUE);
         SetTextColor (hdc, WHITE);
         SetBkMode (hdc, TRANSPARENT);
         sprintf (szText, "%d%%", d);
